Add disclaimer to set public expectations
It can be surprizing to find private links in a public repository,
particularly because they lead to HTTP 404 pages in Github, and may
seem broken links.

There is internal value in keeping this respoitory public, however,
so I want to make sure that no one mistakes that for poor maintenance.