Fix move test suite to latest, stable, Ruby
Ruby 2 currently provides Ruby 2.4.1 (even though 2.5.0 was released),
but 'ruby-head' was providing Ruby 2.6.0dev which is unstable.

The goal of this is to make sure the the gem is usable with a recent Ruby,
we are not aiming at the latest cutting edge version. Whatever Travis CI
considers to be the latest is fine for us.